requirements.txt
asgiref==3.7.2
Django==4.1
djangorestframework==3.14.0
pamqp==2.3.0
pytz==2023.3
rabbitpy==2.0.1
sqlparse==0.4.4
typing_extensions==4.6.3
